backend/requirements.txt
flask==3.0.0
flask-jwt-extended==4.6.0
flask-bcrypt==1.0.1
flask-cors==4.0.0
pillow==10.2.0
